Split out Magnetic Jamming functionality into seperate class
