+++ src/MadgwickAHRS.h
@@ -17,6 +17,7 @@
 #ifndef MadgwickAHRS_h
 #define MadgwickAHRS_h
 #include <math.h>
+#include "MagneticJammingFilter.h"
 
 //--------------------------------------------------------------------------------------------
 // Variable declaration
@@ -28,29 +29,21 @@ private:
     float q1;
     float q2;
     float q3;	// quaternion of sensor frame relative to auxiliary frame
-    float sampleFreq;
     float invSampleFreq;
     float roll;
     float pitch;
     float yaw;
     char anglesComputed;
     void computeAngles();
-    
-    bool firstRun = true;
-    float magMagnitude = NAN;
-    float magMagnitudeFiltered = 42.0f;
-    float magMagnitudeFilteredMax = 57.0f;
-    float magMagnitudeFilteredMin = 37.0f;
-    
-    float magJammingThreshold = 5.0f;
-    bool magJammingActive = false;
-    int magJammingCounter = 0;
 
 //-------------------------------------------------------------------------------------------
 // Function declarations
 public:
     Madgwick(void);
-    void begin(float sampleFrequency) { sampleFreq = sampleFrequency; invSampleFreq = 1.0f / sampleFrequency; }
+    void begin(float sampleFrequency) { 
+		invSampleFreq = 1.0f / sampleFrequency; 
+		magneticJammingFilter.begin(sampleFrequency);
+	}
     void update(float gx, float gy, float gz, float ax, float ay, float az, float mx, float my, float mz);
     void updateIMU(float gx, float gy, float gz, float ax, float ay, float az);
     //float getPitch(){return atan2f(2.0f * q2 * q3 - 2.0f * q0 * q1, 2.0f * q0 * q0 + 2.0f * q3 * q3 - 1.0f);};
@@ -79,23 +72,9 @@ public:
     float getYawRadians() {
         if (!anglesComputed) computeAngles();
         return yaw;
-    }      
-    float getMagneticMagnitude() {
-        return magMagnitude;
-    }
-    float getMagneticMagnitudeFiltered(){
-    	return magMagnitudeFiltered;
-    }
-    bool getMagJammed(){
-    	return magJammingActive;
-	} 
-    void setMagneticJammingThreshold(float magneticJammingThreshold){
-    	if(magneticJammingThreshold < 1.0f)
-    		return;
-    	if(magneticJammingThreshold > 1000.0f)
-    		return;		
-    	magJammingThreshold = magneticJammingThreshold;
-    }    
+    } 
+	    
+    MagneticJammingFilter magneticJammingFilter;
 };
 #endif
